test(onboard): add an in-memory machine runtime fixture
Five machine suites each declared the same in-memory OnboardRuntime
over a session store, and eight duplicated the deep cloneSession
helper byte for byte. Move both into
test/helpers/onboard-machine-runtime-fixture.ts, which also fronts the
session helpers so the suites drop their direct onboard-session imports; the
deterministic clock is overridable where assertions depend on a
specific timestamp, and flow-slices keeps its intentionally inert
markStepFailed and completeSession behavior local. Every migrated
file has a negative line delta.

Refs #8289
